restructer of seeds and types and initial create deposit with test

+++ programs/soljar/src/state/meta_index.rs
@@ -7,7 +7,7 @@ pub struct MetaIndex {
     pub index_page: u32,
     pub created_at: i64,
     pub updated_at: i64,
-    pub total_items: u64,
+    pub total_items: u8,
     #[max_len(50)]
     pub metas: Vec<Pubkey>,
 }
