+++ programs/soljar/src/state/withdrawl_index.rs
@@ -7,7 +7,7 @@ pub struct WithdrawlIndex {
     pub index_page: u32,
     pub created_at: i64,
     pub updated_at: i64,
-    pub total_items: u64,
+    pub total_items: u8,
     #[max_len(50)]
     pub withdrawls: Vec<Pubkey>,
 }
